Added disparity index model, migration, merge w/previous migration. Refactored models so that all stats-type models are in a 'base' module rather than in 'cif'. STATS_MODELS are now registered in the base module rather than cif.
